pyproject.toml
[project]
name = "clinic-registry-backend"
version = "0.1.0"
description = "A backend side for dental clinic registry"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "pre-commit>=4.5.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "python-ulid>=3.1.0",
    "sqlalchemy>=2.0.45",
    "typer>=0.21.0",
    "ulid>=1.1",
    "uvicorn>=0.40.0",
]
